requirements.txt
django==1.10.1
django-contrib-comments==1.9.2
django_mptt==0.11.0
django-tagging==0.5.0
django_blog_zinnia==0.20
django-registration-redux==2.8
psycopg2==2.8.5